+++ sql/ui/eod_settlements.sql
@@ -1,61 +1,43 @@
 
+SELECT * FROM views_asof_date
+WHERE user_id = 'adrian'
+LIMIT 5;
 
-CREATE TABLE IF NOT EXISTS eod_settlements_views (
+
+SELECT DISTINCT user_id, view_name
+FROM views_asof_date
+ORDER BY user_id, view_name;
+
+
+
+---========================================================================
+CREATE TABLE IF NOT EXISTS views_asof_date (
     user_id VARCHAR NOT NULL,
     view_name VARCHAR NOT NULL,
-    source VARCHAR NOT NULL,
     row_id UINTEGER NOT NULL,
+    source VARCHAR NOT NULL,
     ice_category VARCHAR,
     ice_hub VARCHAR,
     ice_product VARCHAR,
     endur_curve_name VARCHAR,
+    nodal_contract_name VARCHAR,
     as_of_date DATE NOT NULL,
     strip VARCHAR,
     unit_conversion VARCHAR,
     label VARCHAR,
 );
 
-INSERT INTO eod_settlements_views 
-VALUES (
-    'adrian', -- user_id
-    'ny capacity', -- view_name
-    'ice', -- source
-    0, -- row_id
-    'power', -- ice_category
-    'NYISO Rest of State Calendar-Month', -- ice_hub
-    'Capacity Futures', -- ice_product
-    NULL, -- endur_curve_name
-    '2026-01-15', -- as_of_date
-    NULL, -- strip
-    NULL, -- unit_conversion
-);
+CREATE TEMPORARY TABLE tmp
+AS 
+    SELECT *
+    FROM read_csv(
+        '/home/adrian/Downloads/Archive/UI/eod_settlements/views_asof_date.csv',
+        header = true,
+        nullstr = ['NULL', 'null', '']
+    )
+    ORDER BY user_id, view_name, row_id;
 
-INSERT INTO eod_settlements_views 
-VALUES (
-    'adrian', -- user_id
-    'ny capacity', -- view_name
-    'ice', -- source
-    1, -- row_id
-    'power', -- ice_category
-    'NYISO Rest of State Calendar-Month', -- ice_hub
-    'Capacity Futures', -- ice_product
-    NULL, -- endur_curve_name
-    '2026-03-20', -- as_of_date
-    NULL, -- strip
-    NULL, -- unit_conversion
-);
+INSERT INTO views_asof_date 
+SELECT *
+FROM tmp;       
 
-INSERT INTO eod_settlements_views 
-VALUES (
-    'adrian', -- user_id
-    'ny capacity', -- view_name
-    'endur', -- source
-    0, -- row_id
-    NULL, -- ice_category
-    NULL, -- ice_hub
-    NULL, -- ice_product
-    'PWR_NYISO_ROS_ICE_MWH_ICAP', -- endur_curve_name
-    '2026-03-20', -- as_of_date
-    NULL, -- strip
-    NULL, -- unit_conversion
-);