+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Move ISONE Monthly Asset NCPC to DuckDB
 - test
 
+# Release 2026-04-15
+- Add ui/eod_settlements client
+
 # Release 2026-03-18
 - Fix urls in the extension methods for LMPs